Reorganize Babbage.Imp module structure
