+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/Imp/UtxosSpec.hs
@@ -1,0 +1,8 @@
+module Test.Cardano.Ledger.Babbage.Imp.UtxosSpec (spec) where
+
+import Test.Cardano.Ledger.Alonzo.ImpTest
+import Test.Cardano.Ledger.Imp.Common
+
+spec :: SpecWith (ImpInit (LedgerSpec era))
+spec = describe "UTXOS" $ do
+  pure ()
